Feature - Average algorithm change
Previously, I would account for all versions of a specific client/platform to calculate the average support of a family. But this implies that, for example, if Apple Mail in iOS 11 didn't support a feature that is now supported in iOS 12, the average would be "Partial". So I now only count the average based on the last version only.